+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/Accels.java
@@ -119,7 +119,7 @@ public class Accels extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			Accels obj = new Accels();
@@ -133,13 +133,13 @@ public class Accels extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public Accels GetInstance(UAVObjectManager objMngr, int instID)
+	public Accels GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (Accels)(objMngr.getObject(Accels.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0xDD9D5FC0;
+	protected static final long OBJID = 0xDD9D5FC0;
 	protected static final String NAME = "Accels";
 	protected static String DESCRIPTION = "The accel data.";
 	protected static final boolean ISSINGLEINST = 1 == 1;
