+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/AltitudeHoldSettings.java
@@ -138,7 +138,7 @@ public class AltitudeHoldSettings extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			AltitudeHoldSettings obj = new AltitudeHoldSettings();
@@ -152,13 +152,13 @@ public class AltitudeHoldSettings extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public AltitudeHoldSettings GetInstance(UAVObjectManager objMngr, int instID)
+	public AltitudeHoldSettings GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (AltitudeHoldSettings)(objMngr.getObject(AltitudeHoldSettings.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0xFEC55B42;
+	protected static final long OBJID = 0xFEC55B42;
 	protected static final String NAME = "AltitudeHoldSettings";
 	protected static String DESCRIPTION = "Settings for the @ref AltitudeHold module";
 	protected static final boolean ISSINGLEINST = 1 == 1;
