+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/FixedWingPathFollowerStatus.java
@@ -133,7 +133,7 @@ public class FixedWingPathFollowerStatus extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			FixedWingPathFollowerStatus obj = new FixedWingPathFollowerStatus();
@@ -147,13 +147,13 @@ public class FixedWingPathFollowerStatus extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public FixedWingPathFollowerStatus GetInstance(UAVObjectManager objMngr, int instID)
+	public FixedWingPathFollowerStatus GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (FixedWingPathFollowerStatus)(objMngr.getObject(FixedWingPathFollowerStatus.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0xA0D6F6D4;
+	protected static final long OBJID = 0xA0D6F6D4;
 	protected static final String NAME = "FixedWingPathFollowerStatus";
 	protected static String DESCRIPTION = "Object Storing Debugging Information on PID internals";
 	protected static final boolean ISSINGLEINST = 1 == 1;
