+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/GyrosBias.java
@@ -115,7 +115,7 @@ public class GyrosBias extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			GyrosBias obj = new GyrosBias();
@@ -129,13 +129,13 @@ public class GyrosBias extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public GyrosBias GetInstance(UAVObjectManager objMngr, int instID)
+	public GyrosBias GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (GyrosBias)(objMngr.getObject(GyrosBias.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0xE4B6F980;
+	protected static final long OBJID = 0xE4B6F980;
 	protected static final String NAME = "GyrosBias";
 	protected static String DESCRIPTION = "The gyro data.";
 	protected static final boolean ISSINGLEINST = 1 == 1;
