+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/PathPlannerSettings.java
@@ -120,7 +120,7 @@ public class PathPlannerSettings extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			PathPlannerSettings obj = new PathPlannerSettings();
@@ -134,13 +134,13 @@ public class PathPlannerSettings extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public PathPlannerSettings GetInstance(UAVObjectManager objMngr, int instID)
+	public PathPlannerSettings GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (PathPlannerSettings)(objMngr.getObject(PathPlannerSettings.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x290E45DA;
+	protected static final long OBJID = 0x290E45DA;
 	protected static final String NAME = "PathPlannerSettings";
 	protected static String DESCRIPTION = "Settings for the @ref PathPlanner Module";
 	protected static final boolean ISSINGLEINST = 1 == 1;
